refactor(PromotionRule): Promotion 에서 PromotionRule 로 변경
