+++ src/test/java/store/domain/promotion/PromotionRuleTest.java
@@ -7,60 +7,26 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class PromotionTest {
+public class PromotionRuleTest {
 
     @Test
     @DisplayName("정상적인 값으로 생성할 수 있다")
     void 프로모션_생성_테스트() {
-        Promotion promotion = Promotion.builder()
-                .name("테스트 프로모션")
+        PromotionRule rule = PromotionRule.builder()
                 .requiredQuantity(3)
                 .freeQuantity(1)
                 .startDate("2024-01-01")
                 .endDate("2024-12-31")
                 .build();
 
-        Assertions.assertThat(promotion.getName()).isEqualTo("테스트 프로모션");
-        Assertions.assertThat(promotion.getRequiredQuantity()).isEqualTo(3);
-        Assertions.assertThat(promotion.getFreeQuantity()).isEqualTo(1);
-    }
-
-    @Test
-    @DisplayName("프로모션명이 null이면 예외가 발생한다")
-    void 프로모션_이름_null_예외가_발생한다() {
-        Assertions.assertThatThrownBy(() -> Promotion.builder()
-                        .name(null)
-                        .requiredQuantity(3)
-                        .freeQuantity(1)
-                        .startDate("2024-01-01")
-                        .endDate("2024-12-31")
-                        .build()
-                )
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 프로모션명은 빈 값일 수 없습니다.");
-    }
-
-    @Test
-    @DisplayName("프로모션명이 빈 문자열이면 예외가 발생한다")
-    void 프로모션_이름_빈_문자열_예외가_발생한다() {
-        Assertions.assertThatThrownBy(() -> Promotion.builder()
-                        .name("  ")
-                        .requiredQuantity(3)
-                        .freeQuantity(1)
-                        .startDate("2024-01-01")
-                        .endDate("2024-12-31")
-                        .build()
-                )
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 프로모션명은 빈 값일 수 없습니다.");
+        Assertions.assertThat(rule).isInstanceOf(PromotionRule.class);
     }
 
     @ParameterizedTest
     @ValueSource(ints = {0, -1})
     @DisplayName("필요 수량이 0 이하면 예외가 발생한다")
     void 필요_수량_0_이하면_예외가_발생한다(int invalidQuantity) {
-        Assertions.assertThatThrownBy(() -> Promotion.builder()
-                        .name("테스트 프로모션")
+        Assertions.assertThatThrownBy(() -> PromotionRule.builder()
                         .requiredQuantity(invalidQuantity)
                         .freeQuantity(1)
                         .startDate("2024-01-01")
@@ -75,8 +41,7 @@ public class PromotionTest {
     @ValueSource(ints = {0, -1})
     @DisplayName("증정 수량이 0 이하면 예외가 발생한다")
     void 증정_수량_0_이하면_예외가_발생한다(int invalidQuantity) {
-        Assertions.assertThatThrownBy(() -> Promotion.builder()
-                        .name("테스트 프로모션")
+        Assertions.assertThatThrownBy(() -> PromotionRule.builder()
                         .requiredQuantity(3)
                         .freeQuantity(invalidQuantity)
                         .startDate("2024-01-01")
@@ -90,8 +55,7 @@ public class PromotionTest {
     @Test
     @DisplayName("시작일이 종료일보다 늦으면 예외가 발생한다")
     void 시작일이_종료일보다_늦으면_예외가_발생한다() {
-        Assertions.assertThatThrownBy(() -> Promotion.builder()
-                        .name("테스트 프로모션")
+        Assertions.assertThatThrownBy(() -> PromotionRule.builder()
                         .requiredQuantity(3)
                         .freeQuantity(1)
                         .startDate("2024-12-31")
@@ -105,8 +69,7 @@ public class PromotionTest {
     @Test
     @DisplayName("프로모션 기간 내의 날짜이면 true를 반환한다")
     void 프로모션_기간_내의_날짜면_true_반환_테스트() {
-        Promotion promotion = Promotion.builder()
-                .name("테스트 프로모션")
+        PromotionRule promotion = PromotionRule.builder()
                 .requiredQuantity(3)
                 .freeQuantity(1)
                 .startDate("2024-01-01")
@@ -119,8 +82,7 @@ public class PromotionTest {
     @Test
     @DisplayName("프로모션 시작일 이전이면 false를 반환한다")
     void 프로모션_시작일_이전이면_false_반환_테스트() {
-        Promotion promotion = Promotion.builder()
-                .name("테스트 프로모션")
+        PromotionRule promotion = PromotionRule.builder()
                 .requiredQuantity(3)
                 .freeQuantity(1)
                 .startDate("2024-01-01")
@@ -133,8 +95,7 @@ public class PromotionTest {
     @Test
     @DisplayName("프로모션 종료일 이후면 false를 반환한다")
     void 프로모션_종료일_이후면_false_반환_테스트() {
-        Promotion promotion = Promotion.builder()
-                .name("테스트 프로모션")
+        PromotionRule promotion = PromotionRule.builder()
                 .requiredQuantity(3)
                 .freeQuantity(1)
                 .startDate("2024-01-01")
